--- FluentValidator.Tests/Domain/Domain.cs
@@ -1,9 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace FluentValidator.Tests.Domain
-{
-    public class Aggregate : Notifiable { }
-    public class Entity : Notifiable { }
-}
